wip: layout topology js refactor
